Add no_std core and alloc feature builds
This should be a purely mechanical change, no API changes. Rawzip
already provides slice reader API and this can be enticing for those
that want a no_std, no alloc, and zero copy zip file reader.
