+++ src/archive.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "std")]
+use crate::Crc32;
 use crate::errors::{Error, ErrorKind};
 use crate::extra_fields::{ExtraFieldId, ExtraFields};
 use crate::headers::EntryFlags;
@@ -6,11 +8,18 @@ use crate::mode::{
     msdos_mode_to_file_mode, unix_mode_to_file_mode,
 };
 use crate::path::{RawPath, ZipFilePath};
-use crate::reader_at::{FileReader, MutexReader, RangeReader, ReaderAt, ReaderAtExt};
+#[cfg(feature = "std")]
+use crate::reader_at::{ReaderAt, ReaderAtExt};
 use crate::time::{DosDateTime, ZipDateTimeKind, extract_best_timestamp};
 use crate::utils::{le_u16, le_u32, le_u64};
-use crate::{Crc32, EndOfCentralDirectory, EndOfCentralDirectoryRecordFixed, ZipLocator};
-use std::io::{Read, Seek, Write};
+use crate::{EndOfCentralDirectory, EndOfCentralDirectoryRecordFixed, ZipLocator};
+#[cfg(feature = "std")]
+use std::io::Write;
+
+#[cfg(feature = "std")]
+mod reader;
+#[cfg(feature = "std")]
+pub use reader::{ZipEntries, ZipEntry, ZipReader, ZipSliceVerifier, ZipVerifier};
 
 pub(crate) const END_OF_CENTRAL_DIR_SIGNATURE64: u32 = 0x06064b50;
 pub(crate) const END_OF_CENTRAL_DIR_LOCATOR_SIGNATURE: u32 = 0x07064b50;
@@ -55,7 +64,7 @@ pub const MAX_CENTRAL_DIRECTORY_RECORD_SIZE: usize =
 ///     println!("Found {} entries.", archive.entries_hint());
 ///     for entry_result in archive.entries() {
 ///         let entry = entry_result?;
-///         println!("File: {}", entry.file_path().try_normalize()?.as_ref());
+///         println!("File: {:?}", entry.file_path().as_ref());
 ///     }
 ///     Ok(())
 /// }
@@ -102,21 +111,33 @@ impl<T: AsRef<[u8]>> ZipSliceArchive<T> {
 
     /// Returns the offset of the End of Central Directory (EOCD) signature.
     ///
-    /// See [`ZipArchive::eocd_offset()`] for more details.
+    /// This is the byte position where the EOCD signature (`0x06054b50`) was
+    /// found. It is useful for recovery scenarios when dealing with false EOCD
+    /// signatures or when restarting archive searches from a known position.
     pub fn eocd_offset(&self) -> u64 {
         self.eocd.tail_eocd_offset()
     }
 
     /// The declared offset of the start of the central directory.
     ///
-    /// See [`ZipArchive::directory_offset()`] for more details.
+    /// To verify the validity of this offset, start iterating through the
+    /// central directory via [`ZipSliceArchive::entries`]. Ensure no errors are
+    /// returned on the first entry.
+    ///
+    /// This value is useful when calculating the amount of prelude data in the
+    /// input, as it serves as the upper bound until each file's
+    /// [`ZipFileHeaderRecord::local_header_offset`] can be examined.
     pub fn directory_offset(&self) -> u64 {
         self.eocd.directory_offset()
     }
 
     /// Returns the offset where the ZIP archive ends.
     ///
-    /// See [`ZipArchive::end_offset`] for more details.
+    /// This returns the position immediately after the last byte of the ZIP
+    /// archive, including the End of Central Directory record and any comment.
+    /// This is useful for extracting trailing data.
+    ///
+    /// The calculation does not rely on self-reported sizes from the archive.
     pub fn end_offset(&self) -> u64 {
         self.eocd.tail_eocd_offset()
             + EndOfCentralDirectoryRecordFixed::SIZE as u64
@@ -142,6 +163,7 @@ impl<T: AsRef<[u8]>> ZipSliceArchive<T> {
     /// This is only needed for `AsRef<[u8]>` types that do not themselves
     /// implement [`ReaderAt`], such as a memory-mapped file (`memmap2::Mmap`).
     /// Otherwise prefer the zero-cost [`ZipSliceArchive::into_reader`].
+    #[cfg(feature = "std")]
     pub fn into_cursor_archive(self) -> ZipArchive<std::io::Cursor<T>> {
         ZipArchive {
             reader: std::io::Cursor::new(self.data),
@@ -151,9 +173,8 @@ impl<T: AsRef<[u8]>> ZipSliceArchive<T> {
 
     /// Seeks to the given file entry in the zip archive.
     ///
-    /// See [`ZipArchive::get_entry`] for more details. The biggest difference
-    /// between the reader and slice APIs is that the slice APIs will eagerly
-    /// validate that the entire compressed data is present.
+    /// The slice API eagerly validates that the entire compressed data is
+    /// present before returning a [`ZipSliceEntry`].
     pub fn get_entry(&self, entry: ZipArchiveEntryWayfinder) -> Result<ZipSliceEntry<'_>, Error> {
         let data = self.data.as_ref();
         let header = &data[(entry.local_header_offset as usize).min(data.len())..];
@@ -207,9 +228,12 @@ impl<'a> ZipSliceEntry<'a> {
         &self.data[self.data_start_offset as usize..]
     }
 
-    /// Returns a verifier for the CRC and uncompressed size of the entry.
+    /// Returns the expected CRC and uncompressed size of the inflated data.
     ///
-    /// See [`ZipReader::claim_verifier`] for more commentary.
+    /// Rawzip considers the central directory to be authoritative, so this
+    /// returns the CRC and uncompressed size from the central directory. If you
+    /// want to check the data descriptor instead, use
+    /// [`ZipSliceEntry::data_descriptor`] to read it.
     pub fn claim_verifier(&self) -> ZipVerification {
         self.verifier
     }
@@ -231,6 +255,7 @@ impl<'a> ZipSliceEntry<'a> {
 
     /// Returns a reader that wraps a decompressor and verify the size and CRC
     /// of the decompressed data once finished.
+    #[cfg(feature = "std")]
     pub fn verifying_reader<D>(&self, reader: D) -> ZipSliceVerifier<D>
     where
         D: std::io::Read,
@@ -245,7 +270,42 @@ impl<'a> ZipSliceEntry<'a> {
 
     /// Returns the byte range of the compressed data within the archive.
     ///
-    /// See [`ZipEntry::compressed_data_range`] for more details.
+    /// This range is calculated from the local file header and identifies the
+    /// compressed data within the original archive bytes.
+    ///
+    /// # Security Usage
+    ///
+    /// This method is useful for detecting overlapping entries, which are often
+    /// used in zip bombs. By comparing the ranges returned by this method
+    /// across multiple entries, you can identify when entries share compressed
+    /// data:
+    ///
+    /// ```rust
+    /// # use rawzip::{ZipArchive, Error};
+    /// # fn example(data: &[u8]) -> Result<(), Error> {
+    /// let archive = ZipArchive::from_slice(data)?;
+    /// let mut ranges = Vec::new();
+    ///
+    /// for entry_result in archive.entries() {
+    ///     let entry = entry_result?;
+    ///     let wayfinder = entry.wayfinder();
+    ///     if let Ok(zip_entry) = archive.get_entry(wayfinder) {
+    ///         ranges.push(zip_entry.compressed_data_range());
+    ///     }
+    /// }
+    ///
+    /// // Check for overlapping ranges
+    /// ranges.sort_by_key(|&(start, _)| start);
+    /// for window in ranges.windows(2) {
+    ///     let (_, end1) = window[0];
+    ///     let (start2, _) = window[1];
+    ///     if end1 > start2 {
+    ///         panic!("Warning: Overlapping entries detected!");
+    ///     }
+    /// }
+    /// # Ok(())
+    /// # }
+    /// ```
     pub fn compressed_data_range(&self) -> (u64, u64) {
         let compressed_data_start = self.local_header_offset + self.data_start_offset as u64;
         let compressed_data_end =
@@ -256,8 +316,7 @@ impl<'a> ZipSliceEntry<'a> {
     /// Returns the local file header information.
     ///
     /// This is the single entry point for the entry's file path, extra fields,
-    /// and flags, mirroring [`ZipEntry::local_header`] on the reader-based
-    /// archive.
+    /// and flags from the local header.
     ///
     /// ```rust
     /// # use rawzip::{ZipArchive, Error};
@@ -289,27 +348,6 @@ impl<'a> ZipSliceEntry<'a> {
             file_path: ZipFilePath::from_bytes(&self.data[filename_start..filename_end]),
             extra_field: &self.data[filename_end..extra_field_end],
         }
-    }
-}
-
-/// Verifies the wrapped reader returns the expected CRC and uncompressed size.
-#[derive(Debug, Clone)]
-pub struct ZipSliceVerifier<Decompressor>(ZipVerifier<Decompressor>);
-
-impl<Decompressor> ZipSliceVerifier<Decompressor> {
-    /// Consumes the [`ZipSliceVerifier`], returning the underlying decompressor
-    /// without verifying.
-    pub fn into_inner(self) -> Decompressor {
-        self.0.into_inner()
-    }
-}
-
-impl<Decompressor> std::io::Read for ZipSliceVerifier<Decompressor>
-where
-    Decompressor: std::io::Read,
-{
-    fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {
-        self.0.read(buf)
     }
 }
 
@@ -363,25 +401,32 @@ impl<'data> Iterator for ZipSliceEntries<'data> {
 
 /// The main entrypoint for reading a Zip archive.
 ///
-/// It can be created from a slice, a file, or any `Read + Seek` source.
-///
-/// # Examples
-///
-/// Creating from a file:
-///
-/// ```rust
-/// # use rawzip::{ZipArchive, Error, RECOMMENDED_BUFFER_SIZE};
-/// # use std::fs::File;
-/// # use std::io;
-/// fn example_from_file(file: File) -> Result<(), Error> {
-///     let mut buffer = vec![0u8; RECOMMENDED_BUFFER_SIZE];
-///     let archive = ZipArchive::from_file(file, &mut buffer)?;
-///     Ok(())
-/// }
-/// ```
+/// It can be created from a slice with [`ZipArchive::from_slice`].
+#[cfg_attr(
+    feature = "std",
+    doc = "With the `std` feature, it can also be created from a file or any `Read + Seek` source."
+)]
 ///
 /// For more complex use cases, use the [`ZipLocator`] to locate an archive.
+#[cfg_attr(
+    feature = "std",
+    doc = r#"
+# Examples
+
+Creating from a file:
+
+```rust
+# fn example_from_file(file: std::fs::File) -> Result<(), rawzip::Error> {
+#     use rawzip::{ZipArchive, RECOMMENDED_BUFFER_SIZE};
+let mut buffer = vec![0u8; RECOMMENDED_BUFFER_SIZE];
+let archive = ZipArchive::from_file(file, &mut buffer)?;
+#     Ok(())
+# }
+```
+"#
+)]
 #[derive(Debug, Clone)]
+#[cfg_attr(not(feature = "std"), allow(dead_code))]
 pub struct ZipArchive<R> {
     reader: R,
     eocd: EndOfCentralDirectory,
@@ -397,439 +442,6 @@ impl ZipArchive<()> {
     /// Parses an archive from in-memory data.
     pub fn from_slice<T: AsRef<[u8]>>(data: T) -> Result<ZipSliceArchive<T>, Error> {
         ZipLocator::new().locate_in_slice(data).map_err(|(_, e)| e)
-    }
-
-    /// Parses an archive from a file by reading the End of Central Directory.
-    ///
-    /// A buffer is required to read parts of the file.
-    /// [`RECOMMENDED_BUFFER_SIZE`] can be used to construct this buffer.
-    pub fn from_file(
-        file: std::fs::File,
-        buffer: &mut [u8],
-    ) -> Result<ZipArchive<FileReader>, Error> {
-        ZipLocator::new()
-            .locate_in_file(file, buffer)
-            .map_err(|(_, e)| e)
-    }
-
-    /// Parses an archive from a seekable reader.
-    ///
-    /// Prefer [`ZipArchive::from_file`] and [`ZipArchive::from_slice`] when
-    /// possible, as they are more efficient due to not wrapping the underlying
-    /// reader in a mutex to support positioned io.
-    ///
-    /// ```rust
-    /// # use rawzip::{ZipArchive, Error, RECOMMENDED_BUFFER_SIZE, ZipFileHeaderRecord};
-    /// # use std::io::Cursor;
-    /// fn example(zip_data: &[u8]) -> Result<(), Error> {
-    ///     let mut buffer = vec![0u8; RECOMMENDED_BUFFER_SIZE];
-    ///     let archive = ZipArchive::from_seekable(Cursor::new(zip_data), &mut buffer)?;
-    ///     Ok(())
-    /// }
-    /// ```
-    pub fn from_seekable<R>(
-        mut reader: R,
-        buffer: &mut [u8],
-    ) -> Result<ZipArchive<MutexReader<R>>, Error>
-    where
-        R: Read + Seek,
-    {
-        let end_offset = reader.seek(std::io::SeekFrom::End(0))?;
-        let reader = MutexReader::new(reader);
-        ZipLocator::new()
-            .locate_in_reader(reader, buffer, end_offset)
-            .map_err(|(_, e)| e)
-    }
-}
-
-impl<R> ZipArchive<R> {
-    pub(crate) fn new(reader: R, eocd: EndOfCentralDirectory) -> Self {
-        ZipArchive { reader, eocd }
-    }
-
-    /// Returns a reference to the underlying reader.
-    pub fn get_ref(&self) -> &R {
-        &self.reader
-    }
-
-    /// Returns a mutable reference to the underlying reader.
-    ///
-    /// This is in contrast to [`ZipSliceArchive`], which cannot safely expose
-    /// mutable access, as it relies on offsets and direct indexing.
-    pub fn get_mut(&mut self) -> &mut R {
-        &mut self.reader
-    }
-
-    /// Consumes this archive and returns the underlying reader.
-    pub fn into_inner(self) -> R {
-        self.reader
-    }
-
-    /// Returns a lending iterator over the entries in the central directory of
-    /// the archive.
-    ///
-    /// Requires a mutable buffer to read directory entries from the underlying
-    /// reader.
-    ///
-    /// ```rust
-    /// # use rawzip::{ZipArchive, Error, RECOMMENDED_BUFFER_SIZE, ZipFileHeaderRecord};
-    /// # use std::fs::File;
-    /// fn example(file: File) -> Result<(), Error> {
-    ///     let mut buffer = vec![0u8; RECOMMENDED_BUFFER_SIZE];
-    ///     let archive = ZipArchive::from_file(file, &mut buffer)?;
-    ///     let entries_hint = archive.entries_hint();
-    ///     let mut actual_entries = 0;
-    ///     let mut entries_iterator = archive.entries(&mut buffer);
-    ///     while let Some(_) = entries_iterator.next_entry()? {
-    ///         actual_entries += 1;
-    ///     }
-    ///     println!("Found {} entries (hint: {})", actual_entries, entries_hint);
-    ///     Ok(())
-    /// }
-    /// ```
-    pub fn entries<'archive, 'buf>(
-        &'archive self,
-        buffer: &'buf mut [u8],
-    ) -> ZipEntries<'archive, 'buf, R> {
-        ZipEntries {
-            buffer,
-            archive: self,
-            pos: 0,
-            end: 0,
-            offset: self.eocd.directory_offset(),
-            base_offset: self.eocd.base_offset(),
-            central_dir_end_pos: self.eocd.head_eocd_offset(),
-        }
-    }
-
-    /// Returns a hint for the total number of entries in the archive.
-    ///
-    /// This value is read from the End of Central Directory record.
-    pub fn entries_hint(&self) -> u64 {
-        self.eocd.entries()
-    }
-
-    /// Returns a Read implementation for the comment of the zip archive.
-    ///
-    /// Use [`RangeReader::remaining()`] to get the comment length before
-    /// reading. It is guaranteed to be less than `u16::MAX`.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use rawzip::{ZipArchive, ZipStr, RECOMMENDED_BUFFER_SIZE};
-    /// use std::io::Read;
-    /// use std::fs::File;
-    ///
-    /// let file = File::open("assets/test.zip")?;
-    /// let mut buffer = vec![0u8; RECOMMENDED_BUFFER_SIZE];
-    /// let archive = ZipArchive::from_file(file, &mut buffer)?;
-    ///
-    /// let mut comment_reader = archive.comment();
-    /// let comment_len = comment_reader.remaining() as usize;
-    /// comment_reader.read_exact(&mut buffer[..comment_len])?;
-    ///
-    /// let actual = ZipStr::new(&buffer[..comment_len]);
-    /// let expected = ZipStr::new(b"This is a zipfile comment.");
-    /// assert_eq!(expected, actual);
-    /// # Ok::<(), Box<dyn std::error::Error>>(())
-    /// ```
-    pub fn comment(&self) -> RangeReader<&R> {
-        let comment_start =
-            self.eocd.tail_eocd_offset() + EndOfCentralDirectoryRecordFixed::SIZE as u64;
-        let comment_end = comment_start + self.eocd.comment_len() as u64;
-        RangeReader::new(&self.reader, comment_start..comment_end)
-    }
-
-    /// Returns the offset of the End of Central Directory (EOCD) signature.
-    ///
-    /// This is the byte position where the EOCD signature (0x06054b50) was found.
-    /// Useful for recovery scenarios when dealing with false EOCD signatures or
-    /// when restarting archive searches from a known position.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # use rawzip::{ZipArchive, ZipLocator, RECOMMENDED_BUFFER_SIZE};
-    /// # use std::fs::File;
-    /// # fn example() -> Result<(), Box<dyn std::error::Error>> {
-    /// # let file = File::open("assets/test.zip")?;
-    /// # let mut buffer = vec![0u8; RECOMMENDED_BUFFER_SIZE];
-    /// let archive = ZipArchive::from_file(file, &mut buffer)?;
-    /// let eocd_position = archive.eocd_offset();
-    ///
-    /// let locator = ZipLocator::new();
-    /// let reader = archive.get_ref();
-    /// let maybe_previous = locator.locate_in_reader(reader, &mut buffer, eocd_position);
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub fn eocd_offset(&self) -> u64 {
-        self.eocd.tail_eocd_offset()
-    }
-
-    /// The declared offset of the start of the central directory.
-    ///
-    /// To verify the validity of this offset, start iterating through the
-    /// central directory via `entries()`. Ensure no errors are returned on the
-    /// first entry.
-    ///
-    /// This value is useful when calculating the amount of prelude data exists
-    /// in the data, as it will serve as the upper bound until each file's
-    /// [`ZipFileHeaderRecord::local_header_offset`] can be examined.
-    pub fn directory_offset(&self) -> u64 {
-        self.eocd.directory_offset()
-    }
-
-    /// Returns the offset where the ZIP archive ends.
-    ///
-    /// This returns the position immediately after the last byte of the ZIP
-    /// archive, including the End of Central Directory record and any comment.
-    /// This is useful for extracting trailing data.
-    ///
-    /// The calculation does not rely on any self reported values from the
-    /// archive.
-    ///
-    /// This can be used in conjunction with the starting offset calculation
-    /// start offset as shown in [`RangeReader`] to determine the exact byte
-    /// range (and thus size) of the ZIP archive within a context of a larger
-    /// file.
-    pub fn end_offset(&self) -> u64 {
-        self.eocd.tail_eocd_offset()
-            + EndOfCentralDirectoryRecordFixed::SIZE as u64
-            + self.comment().remaining()
-    }
-}
-
-impl<R> ZipArchive<R>
-where
-    R: ReaderAt,
-{
-    /// Seeks to the given file entry in the zip archive.
-    pub fn get_entry(&self, entry: ZipArchiveEntryWayfinder) -> Result<ZipEntry<'_, R>, Error> {
-        let mut buffer = [0u8; ZipLocalFileHeaderFixed::SIZE];
-        self.reader
-            .read_exact_at(&mut buffer, entry.local_header_offset)?;
-
-        // The central directory is the source of truth so we really only parse
-        // out the local file header to verify the signature and understand the
-        // variable length. Not everyone uses this as the source of truth:
-        // https://labs.redyops.com/index.php/2020/04/30/spending-a-night-reading-the-zip-file-format-specification/
-        let file_header = ZipLocalFileHeaderFixed::parse(&buffer)?;
-        let (body_offset, o1) = entry
-            .local_header_offset
-            .overflowing_add(ZipLocalFileHeaderFixed::SIZE as u64);
-        let (body_offset, o2) = body_offset.overflowing_add(file_header.variable_length() as u64);
-        let (body_end_offset, o3) = body_offset.overflowing_add(entry.compressed_size);
-
-        if o1 || o2 || o3 {
-            return Err(Error::from(ErrorKind::Eof));
-        }
-
-        Ok(ZipEntry {
-            archive: self,
-            entry,
-            body_offset,
-            body_end_offset,
-        })
-    }
-}
-
-impl<T: ReaderAt> ZipSliceArchive<T> {
-    /// Converts the [`ZipSliceArchive`] into a general [`ZipArchive`].
-    ///
-    /// This is useful for unifying code that might handle both slice-based and
-    /// reader-based archives. Because the underlying data already implements
-    /// [`ReaderAt`], the conversion is zero-cost.
-    pub fn into_reader(self) -> ZipArchive<T> {
-        ZipArchive::from(self)
-    }
-}
-
-impl<R> From<ZipSliceArchive<R>> for ZipArchive<R>
-where
-    R: ReaderAt,
-{
-    fn from(slice_archive: ZipSliceArchive<R>) -> Self {
-        ZipArchive {
-            reader: slice_archive.data,
-            eocd: slice_archive.eocd,
-        }
-    }
-}
-
-/// Represents a single entry (file or directory) within a [`ZipArchive`]
-#[derive(Debug, Clone)]
-pub struct ZipEntry<'archive, R> {
-    archive: &'archive ZipArchive<R>,
-    body_offset: u64,
-    body_end_offset: u64,
-    entry: ZipArchiveEntryWayfinder,
-}
-
-impl<'archive, R> ZipEntry<'archive, R>
-where
-    R: ReaderAt,
-{
-    /// Returns a [`ZipReader`] for reading the compressed data of this entry.
-    pub fn reader(&self) -> ZipReader<&'archive R> {
-        ZipReader {
-            entry: self.entry,
-            range_reader: RangeReader::new(
-                self.archive.get_ref(),
-                self.body_offset..self.body_end_offset,
-            ),
-        }
-    }
-
-    /// Returns a reader that wraps a decompressor and verify the size and CRC
-    /// of the decompressed data once finished.
-    pub fn verifying_reader<D>(&self, reader: D) -> ZipVerifier<D>
-    where
-        D: std::io::Read,
-    {
-        ZipVerifier {
-            reader,
-            crc: Crc32::new(),
-            size: 0,
-            verifier: ZipVerification {
-                crc: self.entry.crc,
-                uncompressed_size: self.entry.uncompressed_size_hint(),
-            },
-        }
-    }
-
-    /// Returns a tuple of start and end byte offsets for the compressed data
-    /// within the underlying reader.
-    ///
-    /// This method uses the information from the local file header in its
-    /// calculations.
-    ///
-    /// # Security Usage
-    ///
-    /// This method is useful for detecting overlapping entries, which are often
-    /// used in zip bombs. By comparing the ranges returned by this method
-    /// across multiple entries, you can identify when entries share compressed
-    /// data:
-    ///
-    /// ```rust
-    /// # use rawzip::{ZipArchive, Error};
-    /// # fn example(data: &[u8]) -> Result<(), Error> {
-    /// let archive = ZipArchive::from_slice(data)?;
-    /// let mut ranges = Vec::new();
-    ///
-    /// for entry_result in archive.entries() {
-    ///     let entry = entry_result?;
-    ///     let wayfinder = entry.wayfinder();
-    ///     if let Ok(zip_entry) = archive.get_entry(wayfinder) {
-    ///         ranges.push(zip_entry.compressed_data_range());
-    ///     }
-    /// }
-    ///
-    /// // Check for overlapping ranges
-    /// ranges.sort_by_key(|&(start, _)| start);
-    /// for window in ranges.windows(2) {
-    ///     let (_, end1) = window[0];
-    ///     let (start2, _) = window[1];
-    ///     if end1 > start2 {
-    ///         panic!("Warning: Overlapping entries detected!");
-    ///     }
-    /// }
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub fn compressed_data_range(&self) -> (u64, u64) {
-        (self.body_offset, self.body_end_offset)
-    }
-
-    /// Returns the local file header information.
-    ///
-    /// This method reads the local file header to which may differ from the
-    /// central directory data. Most ZIP tools use the central directory as
-    /// authoritative, but access to local header data can be useful:
-    ///
-    /// The local header may contain:
-    /// - Additional or different extra fields (richer timestamp data, etc.)
-    /// - Different filename than the central directory (security concern)
-    ///
-    /// The buffer argument must be large enough to hold both the filename and
-    /// extra fields from the local header or a too small error will be
-    /// returned.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # use rawzip::{ZipArchive, RECOMMENDED_BUFFER_SIZE, extra_fields::ExtraFieldId};
-    /// # use std::fs::File;
-    /// # fn example() -> Result<(), Box<dyn std::error::Error>> {
-    /// // Test with filename mismatch test fixture
-    /// let file = File::open("assets/filename_mismatch_test.zip")?;
-    /// let mut buf = vec![0u8; RECOMMENDED_BUFFER_SIZE];
-    /// let archive = ZipArchive::from_file(file, &mut buf)?;
-    ///
-    /// let mut entries = archive.entries(&mut buf);
-    /// let entry_header = entries.next_entry()?.unwrap();
-    ///
-    /// // Central directory shows one filename
-    /// assert_eq!(entry_header.file_path().as_ref(), b"malware.exe");
-    /// let wayfinder = entry_header.wayfinder();
-    /// let entry = archive.get_entry(wayfinder)?;
-    ///
-    /// // Read the local header
-    /// let mut local_buffer = vec![0u8; 1024];
-    /// let local_header = entry.local_header(&mut local_buffer)?;
-    ///
-    /// // Local header shows different filename
-    /// assert_eq!(local_header.file_path().as_ref(), b"safe_file.txt");
-    ///
-    /// // Access extra fields from local header
-    /// let mut found_fields = 0;
-    /// for (field_id, _data) in local_header.extra_fields() {
-    ///     found_fields += 1;
-    ///     // Could check for specific extra field types here
-    ///     println!("Found extra field: {:04x}", field_id.as_u16());
-    /// }
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub fn local_header<'a>(&self, buffer: &'a mut [u8]) -> Result<ZipLocalFileHeader<'a>, Error> {
-        let mut header_buffer = [0u8; ZipLocalFileHeaderFixed::SIZE];
-
-        // Read the local file header
-        self.archive
-            .get_ref()
-            .read_exact_at(&mut header_buffer, self.entry.local_header_offset)?;
-
-        let local_header_fixed =
-            ZipLocalFileHeaderFixed::parse(&header_buffer).expect("header has already been parsed");
-        let file_name_len = local_header_fixed.file_name_len as usize;
-        let extra_field_len = local_header_fixed.extra_field_len as usize;
-        let total_variable_len = file_name_len + extra_field_len;
-
-        // Check if buffer is large enough for both filename and extra fields
-        if buffer.len() < total_variable_len {
-            return Err(Error::from(ErrorKind::BufferTooSmall));
-        }
-
-        let variable_data = &mut buffer[..total_variable_len];
-        let variable_data_offset =
-            self.entry.local_header_offset + ZipLocalFileHeaderFixed::SIZE as u64;
-        self.archive
-            .get_ref()
-            .read_exact_at(variable_data, variable_data_offset)?;
-
-        let (filename_data, extra_field_data) = variable_data.split_at(file_name_len);
-
-        let (compressed_size, uncompressed_size) =
-            local_header_size_hints(&local_header_fixed, extra_field_data);
-
-        Ok(ZipLocalFileHeader {
-            fixed: local_header_fixed,
-            compressed_size,
-            uncompressed_size,
-            file_path: ZipFilePath::from_bytes(filename_data),
-            extra_field: extra_field_data,
-        })
     }
 }
 
@@ -865,102 +477,6 @@ impl ZipVerification {
         }
 
         Ok(())
-    }
-}
-
-/// Verifies the checksum of the decompressed data matches the checksum listed
-/// in the central directory.
-#[derive(Debug, Clone)]
-pub struct ZipVerifier<Decompressor> {
-    reader: Decompressor,
-    crc: Crc32,
-    size: u64,
-    verifier: ZipVerification,
-}
-
-impl<Decompressor> ZipVerifier<Decompressor> {
-    /// Consumes the [`ZipVerifier`], returning the underlying decompressor
-    /// without verifying.
-    pub fn into_inner(self) -> Decompressor {
-        self.reader
-    }
-}
-
-impl<Decompressor> std::io::Read for ZipVerifier<Decompressor>
-where
-    Decompressor: std::io::Read,
-{
-    fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {
-        if buf.is_empty() {
-            return Ok(0);
-        }
-
-        let read = self.reader.read(buf)?;
-        self.crc.update(&buf[..read]);
-        self.size += read as u64;
-
-        if read == 0 || self.size >= self.verifier.uncompressed_size {
-            self.verifier
-                .valid(ZipVerification {
-                    crc: self.crc.checksum(),
-                    uncompressed_size: self.size,
-                })
-                .map_err(|e| std::io::Error::new(std::io::ErrorKind::InvalidData, e))?;
-        }
-
-        Ok(read)
-    }
-}
-
-/// A reader for a Zip entry's compressed data.
-#[derive(Debug, Clone)]
-pub struct ZipReader<R> {
-    entry: ZipArchiveEntryWayfinder,
-    range_reader: RangeReader<R>,
-}
-
-impl<R> ZipReader<R>
-where
-    R: ReaderAt,
-{
-    /// Returns the expected data verification of the inflated data.
-    ///
-    /// Rawzip considers the central directory to be authoritative, so this
-    /// returns the CRC and uncompressed size from the central directory. If you
-    /// want to check the data descriptor instead, use
-    /// [`ZipReader::data_descriptor`] to read it.
-    pub fn claim_verifier(&self) -> ZipVerification {
-        ZipVerification {
-            crc: self.entry.crc,
-            uncompressed_size: self.entry.uncompressed_size_hint(),
-        }
-    }
-
-    /// Reads the trailing [`ZipDataDescriptor`] for this entry, if present.
-    pub fn data_descriptor(&self) -> Result<Option<ZipDataDescriptor>, Error> {
-        if !self.entry.has_data_descriptor {
-            return Ok(None);
-        }
-
-        let descriptor = DataDescriptor::read_at(
-            self.range_reader.get_ref(),
-            self.range_reader.end_offset(),
-            self.entry.data_descriptor_uses_zip64_sizes,
-        )?;
-        Ok(Some(ZipDataDescriptor {
-            crc: descriptor.crc,
-            compressed_size: descriptor.compressed_size,
-            uncompressed_size: descriptor.uncompressed_size,
-        }))
-    }
-}
-
-impl<R> Read for ZipReader<R>
-where
-    R: ReaderAt,
-{
-    fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {
-        self.range_reader.read(buf)
     }
 }
 
@@ -1099,8 +615,7 @@ fn local_header_size_hints(header: &ZipLocalFileHeaderFixed, extra_field: &[u8])
 
 /// The data descriptor trailing an entry's compressed data.
 ///
-/// Obtain one via [`ZipReader::data_descriptor`] or
-/// [`ZipSliceEntry::data_descriptor`].
+/// Obtain one via [`ZipSliceEntry::data_descriptor`].
 ///
 /// The descriptor's size fields are 4 or 8 bytes wide depending on whether the
 /// entry's central directory record signalled zip64 sizes (APPNOTE 4.3.9.2).
@@ -1147,6 +662,7 @@ pub(crate) struct DataDescriptor {
 impl DataDescriptor {
     /// The maximum on-disk size of a data descriptor: optional 4-byte
     /// signature + 4-byte crc + two 8-byte zip64 sizes.
+    #[cfg(feature = "std")]
     const MAX_SIZE: usize = 24;
     pub const SIGNATURE: u32 = 0x08074b50;
 
@@ -1188,6 +704,7 @@ impl DataDescriptor {
         })
     }
 
+    #[cfg(feature = "std")]
     fn read_at<R>(reader: R, offset: u64, uses_zip64_sizes: bool) -> Result<DataDescriptor, Error>
     where
         R: ReaderAt,
@@ -1196,86 +713,6 @@ impl DataDescriptor {
         let mut buffer = [0u8; Self::MAX_SIZE];
         let read = reader.try_read_at_least_at(&mut buffer, Self::MAX_SIZE, offset)?;
         Self::parse(&buffer[..read], uses_zip64_sizes)
-    }
-}
-
-/// A lending iterator over file header records in a [`ZipArchive`].
-#[derive(Debug)]
-pub struct ZipEntries<'archive, 'buf, R> {
-    buffer: &'buf mut [u8],
-    archive: &'archive ZipArchive<R>,
-    pos: usize,
-    end: usize,
-    offset: u64,
-    base_offset: u64,
-    central_dir_end_pos: u64,
-}
-
-impl<R> ZipEntries<'_, '_, R>
-where
-    R: ReaderAt,
-{
-    /// Yield the next zip file entry in the central directory if there is any
-    ///
-    /// This method reads from the underlying archive reader into the provided
-    /// buffer to parse entry headers.
-    #[inline]
-    pub fn next_entry(&mut self) -> Result<Option<ZipFileHeaderRecord<'_>>, Error> {
-        if self.pos + ZipFileHeaderFixed::SIZE >= self.end {
-            if self.offset >= self.central_dir_end_pos {
-                return Ok(None);
-            }
-
-            let remaining = self.end - self.pos;
-            self.buffer.copy_within(self.pos..self.end, 0);
-            let max_read = ((self.central_dir_end_pos - self.offset) as usize)
-                .min(self.buffer.len() - remaining);
-            let read = self.archive.reader.read_at_least_at(
-                &mut self.buffer[remaining..][..max_read],
-                ZipFileHeaderFixed::SIZE,
-                self.offset,
-            )?;
-            self.offset += read as u64;
-            self.pos = 0;
-            self.end = remaining + read;
-        }
-
-        let central_directory_offset = self.offset - (self.end - self.pos) as u64;
-        let data = &self.buffer[self.pos..self.end];
-        let file_header = ZipFileHeaderFixed::parse(data)?;
-        self.pos += ZipFileHeaderFixed::SIZE;
-
-        let variable_length = file_header.variable_length();
-        if self.pos + variable_length > self.end {
-            // Need to read more data
-            let remaining = self.end - self.pos;
-            self.buffer.copy_within(self.pos..self.end, 0);
-            let max_read = ((self.central_dir_end_pos - self.offset) as usize)
-                .min(self.buffer.len() - remaining);
-            let read = self.archive.reader.read_at_least_at(
-                &mut self.buffer[remaining..][..max_read],
-                variable_length - remaining,
-                self.offset,
-            )?;
-            self.offset += read as u64;
-            self.pos = 0;
-            self.end = remaining + read;
-        }
-
-        let data = &self.buffer[self.pos..self.end];
-        let (file_name, extra_field, file_comment, _) = file_header
-            .parse_variable_length(data)
-            .expect("variable length precheck failed");
-        let mut file_header = ZipFileHeaderRecord::from_parts(
-            file_header,
-            file_name,
-            extra_field,
-            file_comment,
-            central_directory_offset,
-        );
-        file_header.local_header_offset += self.base_offset;
-        self.pos += variable_length;
-        Ok(Some(file_header))
     }
 }
 
@@ -1534,6 +971,7 @@ impl<'a> ZipStr<'a> {
 
     /// Converts the borrowed `ZipStr` into an owned `ZipString` by cloning the
     /// data.
+    #[cfg(feature = "alloc")]
     #[inline]
     pub fn into_owned(&self) -> ZipString {
         ZipString::new(self.0.to_vec())
@@ -1543,13 +981,15 @@ impl<'a> ZipStr<'a> {
 /// An owned string (`Vec<u8>`) from a Zip archive, typically for comments or non-path text.
 ///
 /// Similar to `ZipStr`, but owns its data.
+#[cfg(feature = "alloc")]
 #[derive(Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]
-pub struct ZipString(Vec<u8>);
+pub struct ZipString(alloc::vec::Vec<u8>);
 
+#[cfg(feature = "alloc")]
 impl ZipString {
     /// Creates a new `ZipString` from a vector of bytes.
     #[inline]
-    pub fn new(data: Vec<u8>) -> Self {
+    pub fn new(data: alloc::vec::Vec<u8>) -> Self {
         Self(data)
     }
 
@@ -1743,13 +1183,12 @@ impl<'a> ZipFileHeaderRecord<'a> {
     /// The declared offset to the local file header within the Zip archive.
     ///
     /// To verify the validity of this offset, call
-    /// [`ZipSliceArchive::get_entry`] or [`ZipArchive::get_entry`].
+    /// [`ZipSliceArchive::get_entry`].
     ///
     /// The minimum of all local header offsets (or `directory_offset()` when a
     /// zip is empty), will be the length of prelude data in a zip archive (data
     /// that is unrelated to the zip archive).
     ///
-    /// See [`RangeReader`] for an example.
     #[inline]
     pub fn local_header_offset(&self) -> u64 {
         self.local_header_offset
@@ -1773,25 +1212,40 @@ impl<'a> ZipFileHeaderRecord<'a> {
     /// # Example
     /// ```rust
     /// # use rawzip::ZipArchive;
-    /// # fn example() -> Result<(), Box<dyn std::error::Error>> {
+    /// # fn example() -> Result<(), rawzip::Error> {
     /// # let data = include_bytes!("../assets/test.zip");
     /// # let archive = ZipArchive::from_slice(data)?;
     /// # let mut entries = archive.entries();
     /// # let entry = entries.next_entry()?.unwrap();
     /// // Get raw path (potentially unsafe)
     /// let raw_path = entry.file_path();
-    ///
-    /// // Convert to safe path
-    /// let safe_path = raw_path.try_normalize()?;
-    /// println!("Safe path: {}", safe_path.as_ref());
-    ///
-    /// // Check if it's a directory
-    /// if safe_path.is_dir() {
-    ///     println!("This is a directory");
-    /// }
+    /// println!("Raw path bytes: {:?}", raw_path.as_ref());
     /// # Ok(())
     /// # }
+    /// # example()?;
+    /// # Ok::<(), rawzip::Error>(())
     /// ```
+    #[cfg_attr(
+        feature = "alloc",
+        doc = r#"
+With the `alloc` feature, use [`ZipFilePath::try_normalize`] to create a safe path:
+
+```rust
+# use rawzip::ZipArchive;
+# fn example() -> Result<(), Box<dyn std::error::Error>> {
+# let data = include_bytes!("../assets/test.zip");
+# let archive = ZipArchive::from_slice(data)?;
+# let mut entries = archive.entries();
+# let entry = entries.next_entry()?.unwrap();
+let safe_path = entry.file_path().try_normalize()?;
+println!("Safe path: {}", safe_path.as_ref());
+# Ok(())
+# }
+# example()?;
+# Ok::<(), Box<dyn std::error::Error>>(())
+```
+"#
+    )]
     #[inline]
     pub fn file_path(&self) -> ZipFilePath<RawPath<'a>> {
         self.file_name
@@ -1842,9 +1296,9 @@ impl<'a> ZipFileHeaderRecord<'a> {
 
     /// The declared CRC32 checksum of the uncompressed data.
     ///
-    /// To verify the validity of this value, [`ZipEntry::verifying_reader`]
-    /// will return an error if when the decompressed data does not match this
-    /// checksum.
+    /// To verify the validity of this value for slice-backed entries, compare
+    /// it with the [`ZipSliceEntry::claim_verifier`] result while decompressing
+    /// data from [`ZipSliceEntry::data`].
     #[inline]
     pub fn crc32(&self) -> u32 {
         self.crc32
@@ -1937,6 +1391,7 @@ impl ZipArchiveEntryWayfinder {
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
+#[cfg_attr(not(feature = "std"), allow(dead_code))]
 pub(crate) struct ZipLocalFileHeaderFixed {
     pub(crate) signature: u32,
     pub(crate) version_needed: u16,
@@ -1988,6 +1443,7 @@ impl ZipLocalFileHeaderFixed {
         self.file_name_len as usize + self.extra_field_len as usize
     }
 
+    #[cfg(feature = "std")]
     pub fn write<W>(&self, mut writer: W) -> Result<(), Error>
     where
         W: Write,
@@ -2032,6 +1488,7 @@ pub(crate) struct ZipFileHeaderFixed {
 }
 
 impl ZipFileHeaderFixed {
+    #[cfg(feature = "std")]
     pub fn variable_length(&self) -> usize {
         self.file_name_len as usize + self.extra_field_len as usize + self.file_comment_len as usize
     }
@@ -2103,6 +1560,7 @@ impl ZipFileHeaderFixed {
         Some((file_name, extra_field, file_comment, rest))
     }
 
+    #[cfg(feature = "std")]
     pub fn write<W>(&self, mut writer: W) -> Result<(), Error>
     where
         W: Write,
@@ -2134,15 +1592,6 @@ impl ZipFileHeaderFixed {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use std::io::Cursor;
-
-    #[test]
-    pub fn blank_zip_archive() {
-        let data = [80, 75, 5, 6];
-        let mut buf = vec![0u8; RECOMMENDED_BUFFER_SIZE];
-        let archive = ZipArchive::from_seekable(Cursor::new(data), &mut buf);
-        assert!(archive.is_err());
-    }
 
     #[test]
     pub fn trunc_comment_zips() {
@@ -2150,10 +1599,6 @@ mod tests {
             80, 75, 6, 7, 21, 0, 0, 0, 34, 0, 0, 0, 0, 0, 0, 0, 10, 0, 59, 59, 80, 75, 5, 6, 0,
             255, 255, 255, 255, 255, 255, 0, 0, 0, 80, 75, 6, 6, 0, 0, 0, 10,
         ];
-        let mut buf = vec![0u8; RECOMMENDED_BUFFER_SIZE];
-        let archive = ZipArchive::from_seekable(Cursor::new(data), &mut buf);
-        assert!(archive.is_err());
-
         let archive = ZipArchive::from_slice(data);
         assert!(archive.is_err());
     }
@@ -2168,10 +1613,6 @@ mod tests {
 
         let archive = ZipArchive::from_slice(data);
         assert!(archive.is_err());
-
-        let mut buf = vec![0u8; RECOMMENDED_BUFFER_SIZE];
-        let archive = ZipArchive::from_seekable(Cursor::new(data), &mut buf);
-        assert!(archive.is_err());
     }
 
     #[test]
@@ -2184,11 +1625,6 @@ mod tests {
 
         let archive = ZipArchive::from_slice(data).unwrap();
         let mut entries = archive.entries();
-        assert!(entries.next_entry().is_err());
-
-        let mut buf = vec![0u8; RECOMMENDED_BUFFER_SIZE];
-        let archive = ZipArchive::from_seekable(Cursor::new(data), &mut buf).unwrap();
-        let mut entries = archive.entries(&mut buf);
         assert!(entries.next_entry().is_err());
     }
 
@@ -2263,64 +1699,5 @@ mod tests {
         assert_eq!(record.local_header_offset, 0x1_0000_0000);
         assert!(!record.data_descriptor_uses_zip64_sizes);
         assert!(!record.wayfinder().data_descriptor_uses_zip64_sizes);
-    }
-
-    #[test]
-    fn test_compressed_data_range() {
-        let test_zip = std::fs::read("assets/test.zip").unwrap();
-
-        // Test ZipSliceEntry API (from slice)
-        let slice_archive = ZipArchive::from_slice(&test_zip).unwrap();
-        let slice_header_records: Vec<_> = slice_archive
-            .entries()
-            .collect::<Result<Vec<_>, _>>()
-            .unwrap();
-        assert_eq!(slice_header_records.len(), 2);
-
-        let entry1_wayfinder = slice_header_records[0].wayfinder();
-        let slice_entry1 = slice_archive.get_entry(entry1_wayfinder).unwrap();
-        let slice_range1 = slice_entry1.compressed_data_range();
-        assert_eq!(
-            slice_range1,
-            (66, 91),
-            "test.txt compressed data should be at bytes 66-91"
-        );
-
-        let entry2_wayfinder = slice_header_records[1].wayfinder();
-        let slice_entry2 = slice_archive.get_entry(entry2_wayfinder).unwrap();
-        let slice_range2 = slice_entry2.compressed_data_range();
-        assert_eq!(
-            slice_range2,
-            (169, 954),
-            "gophercolor16x16.png compressed data should be at bytes 169-954"
-        );
-
-        let (s1, e1) = slice_range1;
-        assert!(std::ptr::eq(
-            slice_entry1.data(),
-            &test_zip[s1 as usize..e1 as usize]
-        ));
-
-        let (s2, e2) = slice_range2;
-        assert!(std::ptr::eq(
-            slice_entry2.data(),
-            &test_zip[s2 as usize..e2 as usize]
-        ));
-
-        // Test ZipEntry API
-        let file = std::fs::File::open("assets/test.zip").unwrap();
-        let mut buffer = vec![0u8; RECOMMENDED_BUFFER_SIZE];
-        let reader_archive = ZipArchive::from_file(file, &mut buffer).unwrap();
-
-        // Get wayfinders from the slice archive since they should be identical
-        let reader_entry1 = reader_archive.get_entry(entry1_wayfinder).unwrap();
-        let reader_range1 = reader_entry1.compressed_data_range();
-
-        let reader_entry2 = reader_archive.get_entry(entry2_wayfinder).unwrap();
-        let reader_range2 = reader_entry2.compressed_data_range();
-
-        // Verify both APIs return identical ranges
-        assert_eq!(slice_range1, reader_range1);
-        assert_eq!(slice_range2, reader_range2);
     }
 }
